Add pg_tracing extension to contrib
This extension adds the possibility to create server-side spans to be
used for distributed tracing.

For this first version, we rely on SQLCommenter to propagate the trace
context. Other methods of trace propagation (ex: through a GUC variable)
can be added as a follow-up.

The sampling decision to trace a query or not relies on either a global
sampling rate or a sampling rate based on the sampled flag from
SQLCommenter's tracecontext. Additional methods of filtering and tuning
the sampling decision (ex: filtering on queryId) could be added as a
afterward.

When a query is sampled, a top span will be generated using the query
string and used as a parent for the other query's spans. Spans will be
generated for other query functions like Planner, ExecutorRun and
ExecutorFinish.

The spans are exposed through the pg_tracing_peek_spans and
pg_tracing_consume_spans views. Sending spans to a trace collector is
is outside this extension's scope.