Separate memory contexts for relcache and catcache
This patch introduces two new memory contexts under CacheMemoryContext
for relcache and catcache. Most of the time relcache and catcache
constitutes a large part of CacheMemoryContext. Before this patch, it
was hard to see the contribution of both caches in memory usage of
CacheMemoryContext. Having separate contexts
for these two aims to help to understand more about memory usage of
contexts.