Introduce indexes for RestrictInfo
This commit adds indexes to speed up searches for RestrictInfos. When
there are many child partitions and we have to perform planning for join
operations on the tables, we have to handle a large number of
RestrictInfos, resulting in a long planning time.

This commit adds ec_[source|derive]_indexes to speed up the search. We
can use the indexes to filter out unwanted RestrictInfos, and improve
the planning performance.

Author: David Rowley <dgrowley@gmail.com> and me, and includes rebase by
Alena Rybakina <lena.ribackina@yandex.ru> [1].

[1] https://www.postgresql.org/message-id/72d292a1-06ff-432a-a803-af5053786444%40yandex.ru