Allow pg_prewarm to use new smgrreadv
This commit replaces smgrread with smgrreadv (with range).

I also defined a PREWARM_READ_RANGE, 8 by default, which is the
number of buffers to read in one call.

I just discovered pg_readv so I hope the initialization I wrote is correct.

It's not really sure there is a huge win because all blocks are sequential
and readv should merge them but not use vectorization (if I understood
correctly).