Add pg_wait_for_lockers() function
Rather than actually taking any locks on the table(s), it simply waits
for already-held locks to be released, by simply calling the existing
WaitForLockersMultiple() function in the lock manager.

Currently it's only supported on plain tables and doesn't automatically
handle inheritance/partition relationships.

No permissions on the table are required because the same goal can be
achieved (much less efficiently) by repeatedly polling pg_locks, which
doesn't seem to require any permissions. Note that the error messages
reveal the relation name and kind but I think this is OK because these
are visible in the system catalogs?

See email, docs, and tests for more detail.