memory cleaning after DROP VARIABLE
Accept of sinval message invalidates entries in sessionvars hash table. These entries
are validated before any read or write operations on session variables. When the entry
cannot be validated, then it is removed. The removing can be delayed when variable
is dropped inside current transaction. This transaction can be aborted and in this case
we want to preserve content of the variable.