Fix partitionwise join with partially-redundant join clauses
To determine if the two relations being joined can use partitionwise
join, we need to verify the existence of equi-join conditions involving
all the partition keys.  Currently we do that by looking through the
join's restriction clauses.  However, it has been discovered that this
approach is insufficient, because there might be partition keys known
equal by a specific EC, but they do not form a join clause because it
happens that other members of the EC than the partition keys are
constrained to become a join clause.

To address this issue, in addition to examining the join's restriction
clauses, we also check if any partition keys are known equal by ECs, by
leveraging function exprs_known_equal().  To accomplish this, we enhance
exprs_known_equal() to check equality per the semantics of the opfamily,
if provided.