Fix rare recovery shutdown hang due to checkpointer.
Commit 7ff23c6d started running the checkpointer during crash recovery.

As discovered by Justin, in one rare case it could prevent shutdown from
succeeding during a narrow phase at the beginning of crash recovery
after a server crash.

When the the server is automatically restarting but before
PMSIGNAL_RECOVERY_STARTED is received from the startup process,
FatalError is still true.  If a shutdown request arrived in that narrow
window, the PostmasterStateMachine() logic behaved as if the
checkpointer was not running and didn't need to be told to shut down,
and yet waited forever for it to exit.

Now, we can only move from PM_WAIT_BACKENDS state directly to
PM_WAIT_DEADEND if the checkpointer isn't running.  If it is, we now
distinguish between the smart and fast shutdown case where we need to
tell the checkpointer to shutdown and move to PM_SHUTDOWN, and the
immediate shutdown or child crash case where it should already have been
told to quit, and we're still waiting for that to happen so we stay in
PM_WAIT_BACKENDS.

Back-patch to 15.

XXX Experimental patch, not sure yet

Reported-by: Justin Pryzby <pryzby@telsasoft.com>
Discussion: https://postgr.es/m/ZWlrdQarrZvLsgIk@pryzbyj2023