Support absolute bindir/libdir in regression tests with meson
Passing an absolute bindir/libdir will install the binaries and libraries to
<build>/tmp_install/<bindir> and <build>/tmp_install/<libdir> respectively.

This is path is correctly passed to the regression test suite via configure/make,
but not via meson, yet. This is because the "/" operator in the following expression
throws away the whole left side when the right side is an absolute path:

  test_install_location / get_option('libdir')

This was already correctly handled for dir_prefix, which is likely absolute as well.
This patch handles both bindir and libdir in the same way - prefixing absolute paths
with the tmp_install path correctly.