Support falling back to non-preferred readline implementation with meson
To build with -Dreadline=enabled one can use either readline or libedit. The
-Dlibedit_preferred flag is supposed to control the order of names to lookup.
This works fine when either both libraries are present or -Dreadline is set to
auto. However, explicitly enabling readline with only libedit present, but not
setting libedit_preferred, or alternatively enabling readline with only readline
present, but setting libedit_preferred, too, are both broken. This is because
cc.find_library will throw an error for a not found dependency as soon as the
first required dependency is checked, thus it's impossible to fallback to the
alternative.

Here we only check the second of the two dependencies for requiredness, thus
we only fail when none of the two can be found.