Teach predtest.c about BooleanTest
There are lots of things we can prove based on/for IS [NOT]
{TRUE,FALSE,UNKNOWN} clauses -- particularly how they related to boolean
variables, NOT clauses, and IS [NOT] NULL clauses.

There are a few other more generic improvements as well, such as the
ability for clause_is_strict_for to recurse into NOT clauses and
adjusting proofs refutation based on implication to know that if we can
prove strong refutation we can also prove weak refutation by default.