Short-circuit sort_inner_and_outer if there are no mergejoin clauses
In sort_inner_and_outer, we create mergejoin join paths by explicitly
sorting both relations on each possible ordering of the available
mergejoin clauses.  However, if there are no available mergejoin
clauses, we can skip this process entirely.

This patch introduces a check for mergeclause_list at the beginning of
sort_inner_and_outer and exits the function if it is found to be empty.
This might help skip all the statements that come before the call to
select_outer_pathkeys_for_merge, including the build of UniquePath paths
in the case of JOIN_UNIQUE_OUTER or JOIN_UNIQUE_INNER.