POC: serverside SNI support for libpq
Experimental support for serverside SNI support in libpq, a new config
file $datadir/pg_hosts.conf is used for configuring which certicate and
key should be used for which hostname. A new GUC, ssl_snimode, is added
which controls how the hostname TLS extension is handled. The possible
values are off, default and strict:

  - off: pg_hosts.conf is not parsed and the hostname TLS extension is
    not inspected at all. The normal SSL GUCs for certificates and keys
	are used.
  - default: pg_hosts.conf is loaded as well as the normal GUCs. If no
    match for the TLS extension hostname is found in pg_hosts the cert
	and key from the postgresql.conf GUCs is used as the default (used
	as a wildcard host).
  - strict: only pg_hosts.conf is loaded and the TLS extension hostname
    MUST be passed and MUST have a match in the configuration, else the
	connection is refused.

CRL file(s) are applied from postgresql.conf to all configured hostnames.