Fix logging of non-standard pages in RelationCopyStorageUsingBuffer
If the data in the page is non-standard (e.g. the VM, FSM, or an access
method that doesn't update PageHeader's fields as required) then we may
fail to WAL-log the data on that page. When the server crashes after
flushing the WAL record, but before flushing the file writes, this data
may be permanently lost.

Reported-By: Konstantin Knizhnik