Introduce CompactAttribute array in TupleDesc
This array stores a subset of the fields of FormData_pg_attribute,
primarily the ones for deforming tuples, but since we have additional
space, pack a few additional boolean columns in the attflags field.

Many areas of the code can get away with only accessing the
CompactAttribute array, which because the details of each attribute is
stored much more densely than FormData_pg_attribute, many operations can
be performed accessing fewer cachelines which can improve performance.

This also makes pg_attribute.attcacheoff redundant.  A follow-on commit
will remove it.