Check the validity of commutators for merge/hash clauses
When creating merge or hash join plans in createplan.c, the merge or
hash clauses may need to get commuted to ensure that the outer var is on
the left and the inner var is on the right if they are not already in
the expected form.  This requires that their operators have commutators.
Failing to find a commutator at this stage would result in 'ERROR: could
not find commutator for operator xxx', with no opportunity to select an
alternative plan.

Typically, this is not an issue because mergejoinable or hashable
operators are expected to always have valid commutators.  But in some
artificial cases this assumption may not hold true.  Therefore, here in
this patch we check the validity of commutators for clauses in the form
"inner op outer" when selecting mergejoin/hash clauses, and consider a
clause unusable for the current pair of outer and inner relations if it
lacks a commutator.