Avoid unnecessary wrapping for Vars and PHVs
When pulling up a lateral subquery that is underneath an outer join, the
current code always wraps a Var or PHV in the subquery's targetlist into
a new PlaceHolderVar if it is a lateral reference to something outside
the subquery.  This is necessary when the Var/PHV references the
non-nullable side of the outer join from the nullable side: we need to
ensure that it is evaluated at the right place and hence is forced to
null when the outer join should do so.  However, if the referenced rel
is under the same lowest nulling outer join, we can actually omit the
wrapping.  It could be beneficial to get rid of such PHVs because they
imply lateral dependencies, which force us to resort to nestloop joins.

This patch implements this optimization by remembering the relids of the
nullable side of the lowest outer join the subquery is within, and
subsequently checking whether the referenced rel is under the same
lowest nulling outer join.