Use a single GIN tuplesort
The previous approach was to sort the data on a private sort, then read it back,
merge the GinTuples, and write it into the shared sort, to later be used by the
shared tuple sort.

The new approach is to use a single sort, merging tuples as we write them to disk.
This reduces temporary disk space.

An optimization was added to GinBuffer in which we don't deserialize tuples unless
we need access to the itemIds.

This modifies TUplesort to have a new flushwrites callback. Sort's writetup can
now decide to buffer writes until the next flushwrites() callback.