Use pg_pwritev() in XlogWrite()
XlogWrite() had to write() whenever it reached to the last buffer in
circular wal buffers before circling back to the first buffer as
previous pages in wal buffers is not contiguous in memory.

Vectored IO allow us to write when pages are not contiguous. This patch
utilizes pg_pwritev() in XlogWrite() so that it is able to circle back
without having to write() anything if not necessary.