refactor regex related tests
some tests for regexp_instr, regexp_substr using "abcabcabc" as the source string,
which is not good for testing.

Because in many tests the matched pattern is "abc", but in the source string we have 3 "abc"
sometimes we don't know which "abc '' refers to.
changing the source string so the results would be easier to understand.