Fix error handling for non-blocking socket on Windows
The connection fails with a non-blocking socket error when using psql on
Windows to connect to a PostgreSQL server with GSSAPI enabled.

On Windows, the socket error code is obtained by WSAGetLastError() instead of
errno. This causes the value of errno to be incorrect when handling with
non-blocking socket error.

By using the SOCK_ERRNO macro to ensure that all platforms can get the error
code correctly.

Authored-by: Ning Wu <ning94803@gmail.com>