Unwrap a PlaceHolderVar when safe
It's desirable to remove a PlaceHolderVar altogether and use its
contained expression instead when its phnullingrels goes to empty.
Currently we dare not do that because we use PHVs in some cases to
enforce separate identity of subexpressions.  But in cases where the
PHV is used to carry the nullingrel bit of the RTE_GROUP RT index, we
know we can do that.

This patch includes a flag to indicate whether it's safe to unwrap a
PlaceHolderVar, and set this flag to true for the PHVs created in
mark_nullable_by_grouping.