Prevent orphan storage files after server crash
When a server crashes during a transaction that created tables, newly
created but unused storage files are not removed. This patch prevents
such orphan files by utilizing the UNDO log system for storage files.

The behavior of this feature overlaps with the existing functionality,
which handles the removal of unnecessary files during rollback using
pendingDeletes. Therefore, that part will be removed. On the other
hand, the commit-time file deletions are not within the scope of the
UNDO log functionality, so that part will remain used. As a result,
the isCommit flag of the entries in the pendingDeletes list is now
always true. However, to avoid non-essential changes to the code, the
flag will be retained.