Add script to keep .editorconfig in sync with .gitattributes
Our repo already contained an .editorconfig file, but it was not kept up
to date with .gitattributes. This adds a script that keeps these files
in sync. A big advantage of the editorconfig file is that it many
editors/IDEs get automatically configured to trim trailing newlines and
add a final newline on save. While .gitattributes only complains about
these problems instead of automatically fixing them.

This also adds rules to .gitattributes for Python files as well as for
C files in pg_bsd_indent directory (which have a different tab_width
than most C files due to being vendored in).